Fix some refactored classnames
